requirements.txt
## tested python version: 3.12.3
# local 
#--extra-index-url 
# DAS-5
#--extra-index-url https://download.pytorch.org/whl/cu113
# DAS-6
#--extra-index-url https://download.pytorch.org/whl/cu117
# Snellius
#--extra-index-url 

backoff==2.2.1
numpy==1.26.4
matplotlib==3.8.4
torch==2.3.0
torcheval==0.0.7
hydra-core==1.3.2
tqdm==4.66.4
transformers==4.40.2
accelerate==0.31.0
tokenizers==0.19.1
bitsandbytes==0.43.3
openai==1.56.2
regex==2024.5.10
rdflib==7.0.0
mlflow==2.16.2
psutil==5.9.8
pynvml==11.5.0
boto3==1.34.103
ray==2.34.0 
antlr4-python3-runtime==4.9.3
rdflib==7.0.0
stanza==1.8.2
tabulate==0.9.0
nltk==3.8.1
Levenshtein==0.26.0
python-dotenv==1.0.1
sentencepiece==0.2.0
langchain-core==0.3.21
langgraph==0.2.54
# Logic-LM imports
z3-solver==4.13.0.0
ply==3.11